refactor(KERN): remove legacy design autoSuggestInput
