+++ app/components/formElements/autoSuggestInput/KernAutoSuggestClearInput.tsx
@@ -1,10 +1,10 @@
 import type { RefObject } from "react";
 import { useRouteLoaderData } from "react-router";
 import type { ClearIndicatorProps } from "react-select";
+import { KernIcon } from "~/components/kern/common/KernIcon";
 import { type RootLoader } from "~/root";
 import type { DataListOptions } from "~/services/dataListOptions/getDataListOptions";
 import { getTranslationByKey } from "~/services/translations/getTranslationByKey";
-import { KernIcon } from "../../common/KernIcon";
 
 const AUTO_SUGGESTION_INPUT_BUTTON_ARIAL_KEY =
   "auto-suggestion-input-button-arial";
