--- app/components/formElements/autoSuggestInput/customComponents/index.ts
@@ -1,6 +1,0 @@
-export { default as CustomClearIndicator } from "./CustomClearIndicator";
-export { default as CustomControl } from "./CustomControl";
-export { default as CustomValueContainer } from "./CustomValueContainer";
-export { default as CustomInput } from "./CustomInput";
-export { default as customStyles } from "./customStyles";
-export { default as FormatOptionLabel } from "./FormatOptionLabel";
