+++ app/components/formElements/schemaToForm/renderZodString.tsx
@@ -8,8 +8,8 @@ import TextInput, {
 import NumberInput from "~/components/kern/formElements/input/NumberInput";
 import TelephoneInput from "~/components/kern/formElements/input/TelephoneInput";
 import KernTimeInput from "~/components/kern/formElements/input/KernTimeInput";
-import KernAutoSuggestInput from "~/components/kern/formElements/autoSuggest/KernAutoSuggestInput";
 import KernDateInput from "~/components/formElements/inputs/KernDateInput";
+import KernAutoSuggestInput from "../autoSuggestInput/KernAutoSuggestInput";
 
 export const isZodString = (
   fieldSchema: z.ZodType,
