+++ stories/form/AutoSuggestionInput.stories.tsx
@@ -1,9 +1,9 @@
 import { reactRouterFormContext } from "~/../.storybook/reactRouterFormContext";
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import KernAutoSuggestInput from "~/components/kern/formElements/autoSuggest/KernAutoSuggestInput";
 import { Grid } from "~/components/layout/grid/Grid";
 import { GridItem } from "~/components/layout/grid/GridItem";
 import { GridSection } from "~/components/layout/grid/GridSection";
+import KernAutoSuggestInput from "~/components/formElements/autoSuggestInput/KernAutoSuggestInput";
 
 const meta = {
   title: "form/AutoSuggestionInput",
